+++ EmojidexAndroidLibrary/src/main/java/com/emojidex/emojidexandroid/Emoji.java
@@ -6,6 +6,7 @@ import android.graphics.Bitmap;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
 
+import com.emojidex.emojidexandroid.animation.EmojidexAnimationDrawable;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
